+++ include/ui.h
@@ -7,9 +7,9 @@
 
 struct UI
 {
-    uint8_t circle_r = 5;
-    uint8_t circle_x = OLED_WIDTH / 2;
-    uint8_t circle_y = OLED_HEIGHT - circle_r - 5;
+    uint16_t circle_r = 5;
+    uint16_t circle_x = OLED_WIDTH / 2;
+    uint16_t circle_y = OLED_HEIGHT - circle_r - 5;
 
     void init();
     void update_frame();
